fix bug

+++ src/main/java/com/zes/squad/gmh/web/entity/po/ConsumeRecordPo.java
@@ -25,7 +25,7 @@ public class ConsumeRecordPo extends Po {
     private String            consumerName;         //可以是会员也可以是非会员
     private BigDecimal        charge;
     private Integer           chargeWay;
-    private Integer           counselor;
+    private Long              counselor;
     private String            source;
     private Date              consumeTime;
     private String            remark;
